pyproject.toml
[project]
name = "thrive-ai"
version = "0"
description = "Thrive AI Chatbot"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "anthropic>=0.49.0",
    "chromadb>=0.6.3",
    "ethical-guardrails-lib",
    "kaleido==0.2.1",
    "ollama>=0.4.7",
    "openai>=1.66.3",
    "psycopg2-binary>=2.9.10",
    "pyaudio>=0.2.14",
    "requests>=2.31.0",
    "pyperclip>=1.9.0",
    "pyttsx3>=2.98",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "sketch>=0.5.2",
    "speechrecognition>=3.14.1",
    "streamlit>=1.43.2",
    "streamlit-cookies-manager-ext>=0.1.0",
    "vanna[anthropic,chromadb,gemini,ollama,postgres]==0.7.5",
    "wordcloud>=1.9.4",
    "vertexai>=1.71.1",
    "pymilvus>=2.6.0; sys_platform == 'linux' or sys_platform == 'darwin'",
    "milvus-lite>=2.5.1; sys_platform == 'linux' or sys_platform == 'darwin'",
    "openpyxl>=3.1.0",
    "pydantic-ai>=1.0.8",
    "docling>=2.65.0",
    "typer>=0.15.0",
    "rich>=13.0.0",
    "pydantic-settings>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.2.1",
    "ruff>=0.11.9",
    "ty>=0.0.4",
    "watchdog>=6.0.0",
]

[tool.ruff]
# Allow lines to be as long as 120.
line-length = 120

[tool.uv.sources]
ethical-guardrails-lib = { git = "https://github.com/ThriveAI-Solutions/ethical_guardrails_lib.git" }
